Merging stable version 3

+++ Assets/1_Content/Scripts/Runtime/Entities/Components/Weapon/GeneralWeaponMod.cs
@@ -7,7 +7,7 @@ namespace BH.Runtime.Entities
     public class GeneralWeaponMod
     {
         [field: SerializeField]
-        public float IncreasedDamage { get; set; } = 0;
+        public int IncreasedDamage { get; set; } = 0;
 
         [field: SerializeField]
         public float DamageMultiplier { get; set; } = 1;
